# compose_api/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: compose-api
  version: 0.7.1
paths:
...
  /simulations/{simulation_id}:
    get:
      tags:
      - Simulations
      summary: 'Describe one simulation: its latest SLURM job, and how many events
        and datasets its run recorded'
      operationId: get-simulation
      parameters:
      - name: simulation_id
        in: path
        required: true
        schema:
          type: integer
          title: Simulation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationDetail'
        '404':
          description: The simulation does not exist, or the caller may not read it
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /results/simulations/status/batch:
    get:
      tags:
      - Results
      summary: Get simulation status records for a list of IDs
      operationId: get-simulations-status-batch
      requestBody:
        content:
          application/json:
            schema:
              items:
                type: integer
              type: array
              title: Ids
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/HpcRun'
                type: array
                title: Response Get-Simulations-Status-Batch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /results/simulation/status:
    get:
      tags:
      - Results
      summary: Get the simulation status record by its ID
      operationId: get-simulation-status
      parameters:
      - name: simulation_id
        in: query
        required: true
        schema:
          type: integer
          title: Simulation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HpcRun'
        '404':
          description: The requested resource does not exist
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /results/simulation/results/file:
    get:
      tags:
      - Results
      summary: Get simulation results as a zip file
      operationId: get-simulation-results-file
      parameters:
      - name: simulation_id
        in: query
        required: true
        schema:
          type: integer
          title: Simulation Id
      responses:
        '200':
          description: Simulation result zip file
          content:
            application/octet-stream:
              schema:
                format: binary
        '404':
          description: The simulation does not exist, or its results are not available
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /results/simulator/build/status:
    get:
      tags:
      - Results
      summary: Get the simulator build status record by its ID
      operationId: get-simulator-build-status
      parameters:
      - name: simulator_id
        in: query
        required: true
        schema:
          type: integer
          title: Simulator Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HpcRun'
        '404':
          description: The requested resource does not exist
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /results/simulation/events:
    get:
      tags:
      - Results
      summary: Get a page of a simulation's events, in the order they were recorded
      operationId: get-simulation-events
      parameters:
      - name: after
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Return events after this cursor (a previous page's next)
          title: After
        description: Return events after this cursor (a previous page's next)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 5000
          minimum: 1
          default: 500
          title: Limit
      - name: level
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Only events at this level (debug, info, warning, error)
          title: Level
        description: Only events at this level (debug, info, warning, error)
      - name: event
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Only events with this name, e.g. job.end
          title: Event
        description: Only events with this name, e.g. job.end
      - name: span_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Only events inside this span
          title: Span Id
        description: Only events inside this span
      - name: simulation_id
        in: query
        required: true
        schema:
          type: integer
          title: Simulation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEventPage'
        '404':
          description: The requested resource does not exist
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /results/simulation/trace:
    get:
      tags:
      - Results
      summary: Get a simulation's spans as a tree, each span with its own events
      operationId: get-simulation-trace
      parameters:
      - name: simulation_id
        in: query
        required: true
        schema:
          type: integer
          title: Simulation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunTraceTree'
        '404':
          description: The requested resource does not exist
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /results/simulation/trace/chrome:
    get:
      tags:
      - Results
      summary: Get a simulation's trace as a Chrome Trace Event document (open it
        in ui.perfetto.dev)
      operationId: get-simulation-trace-chrome
      parameters:
      - name: simulation_id
        in: query
        required: true
        schema:
          type: integer
          title: Simulation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get-Simulation-Trace-Chrome
        '404':
          description: The requested resource does not exist
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HpcRun:
      properties:
        database_id:
          type: integer
          title: Database Id
        slurmjobid:
          type: integer
          title: Slurmjobid
        correlation_id:
          type: string
          title: Correlation Id
        job_type:
          $ref: '#/components/schemas/JobType'
        sim_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sim Id
        simulator_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Simulator Id
        status:
          anyOf:
          - $ref: '#/components/schemas/JobStatus'
          - type: 'null'
        start_time:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
          - type: 'null'
          title: End Time
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
      type: object
      required:
      - database_id
      - slurmjobid
      - correlation_id
      - job_type
      - sim_id
      - simulator_id
      title: HpcRun
    JobStatus:
      type: string
      enum:
      - waiting
      - queued
      - running
      - completed
      - failed
      - pending
      - cancelled
      - out_of_memory
      - suspended
      - timeout
      - unknown
      title: JobStatus
    JobType:
      type: string
      enum:
      - simulation
      - build_container
      title: JobType
...
    RunEvent:
      properties:
        cursor:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cursor
        seq:
          type: integer
          title: Seq
        source:
          type: string
          title: Source
        ts:
          type: string
          title: Ts
        component:
          type: string
          title: Component
        event:
          type: string
          title: Event
        level:
          type: string
          title: Level
          default: info
        baggage:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Baggage
        global_time:
          anyOf:
          - type: number
          - type: 'null'
          title: Global Time
        wall_time:
          anyOf:
          - type: number
          - type: 'null'
          title: Wall Time
        span_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Span Id
        parent_span_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Span Id
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
        tags:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Tags
      type: object
      required:
      - seq
      - source
      - ts
      - component
      - event
      title: RunEvent
      description: One structured event of a run, from the API, the job script or
        the simulator's engine.
    RunEventPage:
      properties:
        simulation_id:
          type: integer
          title: Simulation Id
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        events:
          items:
            $ref: '#/components/schemas/RunEvent'
          type: array
          title: Events
        next_cursor:
          anyOf:
          - type: integer
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - simulation_id
      - trace_id
      - events
      title: RunEventPage
      description: 'A page of a simulation''s events, in the order they were recorded.


        ``next_cursor`` is the ``after`` for the next page, None at the end.'
    RunSpan:
      properties:
        span_id:
          type: string
          title: Span Id
        parent_span_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Span Id
        name:
          type: string
          title: Name
        attrs:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Attrs
        start_ts:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Ts
        end_ts:
          anyOf:
          - type: string
          - type: 'null'
          title: End Ts
        duration_s:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration S
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - span_id
      - name
      title: RunSpan
      description: A node of a run's trace tree, folded from ``span.start`` / ``span.end``
        events. ``end_ts`` is None while open.
    RunTraceTree:
      properties:
        simulation_id:
          type: integer
          title: Simulation Id
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        roots:
          items:
            $ref: '#/components/schemas/SpanTree'
          type: array
          title: Roots
      type: object
      required:
      - simulation_id
      - trace_id
      - roots
      title: RunTraceTree
      description: A simulation's spans as a forest, each span carrying its own events.
    SimulationDetail:
      properties:
        simulation_id:
          type: integer
          title: Simulation Id
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        experiment_id:
          type: string
          title: Experiment Id
        simulator_id:
          type: integer
          title: Simulator Id
        simulator:
          anyOf:
          - type: string
          - type: 'null'
          title: Simulator
        container_def_hash:
          type: string
          title: Container Def Hash
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: public
        status:
          type: string
          title: Status
        slurm_job_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Slurm Job Id
        start_time:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
          - type: 'null'
          title: End Time
        exit_code:
          anyOf:
          - type: integer
          - type: 'null'
          title: Exit Code
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        event_count:
          type: integer
          title: Event Count
          default: 0
        dataset_count:
          type: integer
          title: Dataset Count
          default: 0
      type: object
      required:
      - simulation_id
      - experiment_id
      - simulator_id
      - container_def_hash
      - status
      title: SimulationDetail
      description: A simulation with what its run recorded.
...
    SpanTree:
      properties:
        span:
          $ref: '#/components/schemas/RunSpan'
        events:
          items:
            $ref: '#/components/schemas/RunEvent'
          type: array
          title: Events
        children:
          items:
            $ref: '#/components/schemas/SpanTree'
          type: array
          title: Children
      type: object
      required:
      - span
      title: SpanTree
      description: A span with its own events (those whose ``span_id`` is the span's)
        and its children, by start time.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Visibility:
      type: string
      enum:
      - public
      - private
      title: Visibility
      description: Who may read a simulation and everything it produced (plan-observability
        O7).
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional Auth0 access token for this API. Omit it to call anonymously.
security:
- {}
- BearerAuth: []